Store Addy API key in standard config file location
Rather than storing the Addy API key in the same directory as the
Python package source file that reads it, store it in the user config
directory, using the `platformdirs` package to determine where that
is.

The old location (where the source file is) is supported for reading
the key for backward compatibility, but it should eventually be
removed.